Move tracing setup to the `tracing` module
